+++ NEWS.md
@@ -1,5 +1,8 @@
 # fdistr 0.1.0.9000
 
+* Added `create_ecdf` function to generate empirical cumulative distribution
+  function plots.
+
 # fdistr 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.